Temporary fix for ZFS memory computations on Solaris systems
In low-memory scenerios, our computation for available memory is
clearly incorrect. Currently, available memory is computed as:
    (Available Pages * page size) + ZFS_cache_size
But, in low memory scenerios, this can work out to be greater than
the total memory on the system, causing problems down the road.

For now, if adding ZFS_cache_size goes over our total memory, we
will ignore the ZFS cache.

A call to Oracle Technical Support is outstanding to better understand
this scenerio.